Create labels view for reuse
